requirements.txt
selenium==4.16.0
beautifulsoup4==4.12.2
requests==2.31.0
paho-mqtt==1.6.1
